process_and_embed/requirements.txt

# requirements.txt for process-and-embed (with Firestore)
functions-framework==3.4.0
google-cloud-storage==2.16.0
requests==2.31.0
google-auth==2.29.0
google-cloud-firestore==2.16.0